+++ libstreaming/src/main/java/net/majorkernelpanic/streaming/audio/AudioDataListener.java
@@ -1,5 +1,5 @@
 package net.majorkernelpanic.streaming.audio;
 
 public interface AudioDataListener {
-    void onDataReady(byte[] data);
+    void onDataReady(short[] data);
 }
